Fixed #ifdef checks that were using i386 to use __i386__
"""
It's simply obsolete, sloppy, compiler namespace pollution.  The
compiler is not allowed to predefine symbols that might conflict with
ordinary identifiers.  For backwards compatibility gcc currently
predefines i386 when compiling for x86 32-bit (but not 64-bit), but that
will go away.  It is also not defined if you specify -ansi when invoking
the compiler, because then it is seriously standards compliant.  Other
compilers shouldn't define it either.  Correct code shouldn't rely on it
being defined.  However __i386__ is safe and proper.
"""